Fixed code format issues
