+++ roadrescue/screens/EditProfile.js
@@ -1,12 +1,12 @@
 import React, {Component} from 'react';
 import {
-  StyleSheet,
   View,
   Text,
   TouchableOpacity,
   KeyboardAvoidingView,
   Image,
 } from 'react-native';
+import {styles} from './styles/Style';
 import user from '../assets/images/user.png';
 import Icon from 'react-native-vector-icons/FontAwesome';
 import {SafeAreaView} from 'react-navigation';
@@ -40,18 +40,14 @@ class EditProfile extends Component {
           <ScrollView>
             <KeyboardAvoidingView style={styles.wrapper}>
               <View style={styles.cardStyle}>
-                <Image style={styles.image} source={user} />
+                <Image style={styles.editImage} source={user} />
               </View>
 
               <Text style={styles.conText}>Edit your profile</Text>
 
               <View style={[styles.card2]}>
-                <View style={styles.container}>
-                  <Icon
-                    name="user"
-                    size={30}
-                    style={{padding: 10, marginTop: 10, color: '#fff'}}
-                  />
+                <View style={styles.madokaContainer}>
+                  <Icon name="user" size={30} style={styles.ImageStyle} />
                   <Madoka
                     style={styles.inputFlex2}
                     label={'Full name'}
@@ -60,12 +56,8 @@ class EditProfile extends Component {
                     inputStyle={{color: '#fff'}}
                   />
                 </View>
-                <View style={styles.container}>
-                  <Icon
-                    name="envelope"
-                    size={30}
-                    style={{padding: 10, marginTop: 10, color: '#fff'}}
-                  />
+                <View style={styles.madokaContainer}>
+                  <Icon name="envelope" size={30} style={styles.ImageStyle} />
                   <Madoka
                     style={styles.inputFlex2}
                     label={'email'}
@@ -74,12 +66,8 @@ class EditProfile extends Component {
                     inputStyle={{color: '#fff'}}
                   />
                 </View>
-                <View style={styles.container}>
-                  <Icon
-                    name="phone"
-                    size={30}
-                    style={{padding: 10, marginTop: 10, color: '#fff'}}
-                  />
+                <View style={styles.madokaContainer}>
+                  <Icon name="phone" size={30} style={styles.ImageStyle} />
                   <Madoka
                     style={styles.inputFlex2}
                     label={'phone'}
@@ -88,12 +76,8 @@ class EditProfile extends Component {
                     inputStyle={{color: '#fff'}}
                   />
                 </View>
-                <View style={styles.container}>
-                  <Icon
-                    name="map-marker"
-                    size={30}
-                    style={{padding: 10, marginTop: 10, color: '#fff'}}
-                  />
+                <View style={styles.madokaContainer}>
+                  <Icon name="map-marker" size={30} style={styles.ImageStyle} />
                   <Madoka
                     style={styles.inputFlex2}
                     label={'address'}
@@ -103,19 +87,15 @@ class EditProfile extends Component {
                   />
                 </View>
 
-                <View style={styles.container}>
-                  <Icon
-                    name="lock"
-                    size={30}
-                    style={{padding: 10, marginTop: 10, color: '#fff'}}
-                  />
+                <View style={styles.madokaContainer}>
+                  <Icon name="lock" size={30} style={styles.ImageStyle} />
                   <Madoka
                     secureTextEntry={this.state.hidePassword}
                     style={styles.inputFlex2}
                     label={'password'}
                     borderColor={'#fff'}
-                    labelStyle={{color: '#fff'}}
-                    inputStyle={{color: '#fff'}}
+                    labelStyle={styles.labelColor}
+                    inputStyle={styles.labelColor}
                   />
                   <TouchableOpacity
                     activeOpacity={0.8}
@@ -143,119 +123,5 @@ class EditProfile extends Component {
     );
   }
 }
-
-const styles = StyleSheet.create({
-  wrapper: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingLeft: 40,
-    paddingRight: 40,
-  },
-  container: {
-    flex: 1,
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    justifyContent: 'space-around',
-  },
-  safeArea: {
-    flex: 1,
-  },
-  conText: {
-    color: '#fff',
-    fontSize: 17,
-    fontWeight: 'bold',
-    marginTop: 20,
-  },
-  logText: {
-    color: '#fff',
-    fontSize: 15,
-    fontWeight: 'bold',
-    marginTop: 20,
-  },
-  loginText: {
-    color: '#fff',
-    fontSize: 15,
-    fontWeight: 'bold',
-    textTransform: 'uppercase',
-  },
-  ImageStyle: {
-    padding: 10,
-    margin: 5,
-    height: 25,
-    width: 25,
-    resizeMode: 'stretch',
-    alignItems: 'center',
-  },
-  btn: {
-    alignSelf: 'stretch',
-    alignItems: 'center',
-    padding: 20,
-    marginTop: 20,
-    borderRadius: 25,
-    backgroundColor: '#A40606',
-  },
-  image: {
-    width: 50,
-    height: 50,
-    marginTop: 20,
-    padding: 20,
-  },
-  content: {
-    paddingBottom: 300,
-  },
-  card1: {
-    paddingVertical: 16,
-  },
-  card2: {
-    padding: 14,
-    alignSelf: 'stretch',
-    marginTop: 20,
-  },
-  input: {
-    marginTop: 4,
-  },
-  inputFlex: {
-    marginTop: 4,
-    alignSelf: 'stretch',
-    width: 150,
-  },
-  inputFlex2: {
-    alignSelf: 'stretch',
-    width: 250,
-  },
-  title: {
-    paddingBottom: 16,
-    textAlign: 'center',
-    color: '#404d5b',
-    fontSize: 20,
-    fontWeight: 'bold',
-    opacity: 0.8,
-  },
-  cardStyle: {
-    flex: 1,
-    alignItems: 'center',
-    width: 100,
-    height: 100,
-    borderTopLeftRadius: 100,
-    borderTopRightRadius: 100,
-    borderBottomLeftRadius: 100,
-    borderBottomRightRadius: 100,
-    backgroundColor: '#fff',
-  },
-  visibilityBtn: {
-    position: 'absolute',
-    right: 15,
-    height: 40,
-    width: 35,
-    padding: 3,
-  },
-  btnImage: {
-    resizeMode: 'contain',
-    height: '100%',
-    width: '100%',
-    marginTop: 7,
-  },
-});
 
 export default EditProfile;
