+++ roadrescue/screens/styles/Style.js
@@ -1,0 +1,212 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    height: 150,
+    padding: 10,
+    borderTopLeftRadius: 5,
+    borderTopRightRadius: 5,
+    borderBottomLeftRadius: 5,
+    borderBottomRightRadius: 5,
+    backgroundColor: '#F5FFFA',
+  },
+  homeContainer: {
+    backgroundColor: '#8c231a',
+    padding: 10,
+  },
+  containerText: {
+    color: '#fff',
+    fontSize: 20,
+    fontWeight: 'bold',
+  },
+  menuIcon: {
+    alignSelf: 'flex-start',
+    marginTop: -10,
+    margin: 18,
+  },
+  crashText: {
+    textAlign: 'center',
+    color: '#fff',
+    fontWeight: 'bold',
+    fontSize: 18,
+    borderTopLeftRadius: 5,
+    borderTopRightRadius: 5,
+    borderBottomLeftRadius: 5,
+    borderBottomRightRadius: 5,
+  },
+  subHomeContainer: {
+    alignSelf: 'center',
+    marginTop: 10,
+    padding: 44,
+  },
+  imageIcon: {
+    alignSelf: 'center',
+    width: 70,
+    height: 70,
+  },
+  buttonTouch: {
+    alignSelf: 'center',
+    marginTop: 10,
+    backgroundColor: '#8c231a',
+    width: 150,
+    padding: 10,
+  },
+  menuContainer: {
+    alignItems: 'flex-start',
+    margin: 18,
+  },
+  aboutContainer: {
+    flex: 1,
+  },
+  madokaContainer: {
+    flex: 1,
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    justifyContent: 'space-around',
+    marginTop: 10,
+  },
+  menuItem: {
+    alignItems: 'flex-start',
+    margin: 18,
+  },
+  madokaStyle: {
+    color: '#fff',
+  },
+  wrapper: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    paddingLeft: 40,
+    paddingRight: 40,
+  },
+  safeArea: {
+    flex: 1,
+    marginTop: 40,
+  },
+  editImage: {
+    width: 50,
+    height: 50,
+    marginTop: 20,
+    padding: 20,
+  },
+    labelColor:{
+    color: '#fff',
+  },
+  textInput: {
+    alignSelf: 'stretch',
+    borderColor: '#36c',
+    height: 45,
+    borderWidth: 1,
+    borderBottomWidth: 1,
+    borderTopLeftRadius: 1,
+    borderTopRightRadius: 1,
+    borderBottomRightRadius: 1,
+    borderBottomLeftRadius: 1,
+    padding: 10,
+    marginLeft: 50,
+    fontSize: 16,
+  },
+  conText: {
+    color: '#fff',
+    fontSize: 17,
+    fontWeight: 'bold',
+    marginTop: 20,
+  },
+  msgBox: {
+    alignSelf: 'stretch',
+    borderColor: '#fff',
+    color: '#fff',
+    height: 130,
+    borderWidth: 2,
+    borderTopLeftRadius: 2,
+    borderTopRightRadius: 2,
+    borderBottomRightRadius: 2,
+    borderBottomLeftRadius: 2,
+    marginTop: 10,
+    padding: 20,
+    marginStart: 20,
+    textAlignVertical: 'top',
+  },
+  loginText: {
+    color: '#fff',
+    fontSize: 15,
+    fontWeight: 'bold',
+    textTransform: 'uppercase',
+  },
+  ImageStyle: {
+    padding: 10,
+    marginTop: 10,
+    color: '#fff',
+  },
+  btn: {
+    alignSelf: 'stretch',
+    alignItems: 'center',
+    padding: 20,
+    marginTop: 20,
+    borderRadius: 25,
+    backgroundColor: '#A40606',
+  },
+  card2: {
+    padding: 10,
+    alignSelf: 'stretch',
+    marginTop: 20,
+  },
+  inputFlex2: {
+    alignSelf: 'stretch',
+    width: 250,
+  },
+  title: {
+    paddingBottom: 16,
+    textAlign: 'center',
+    color: '#404d5b',
+    fontSize: 20,
+    fontWeight: 'bold',
+    opacity: 0.8,
+  },
+  cardStyle: {
+    flex: 1,
+    alignItems: 'center',
+    width: 100,
+    height: 100,
+    borderTopLeftRadius: 100,
+    borderTopRightRadius: 100,
+    borderBottomLeftRadius: 100,
+    borderBottomRightRadius: 100,
+    backgroundColor: '#fff',
+  },
+  logText: {
+    color: '#fff',
+    fontSize: 15,
+    fontWeight: 'bold',
+    marginTop: 20,
+  },
+  contactImage: {
+    width: 80,
+    height: 70,
+    marginTop: 20,
+    padding: 10,
+  },
+  title: {
+    paddingBottom: 16,
+    textAlign: 'center',
+    color: '#404d5b',
+    fontSize: 20,
+    fontWeight: 'bold',
+    opacity: 0.8,
+  },
+  visibilityBtn: {
+    position: 'absolute',
+    right: 15,
+    height: 40,
+    width: 35,
+    padding: 3,
+  },
+  btnImage: {
+    resizeMode: 'contain',
+    height: '100%',
+    width: '100%',
+    marginTop: 7,
+  },
+});
+
+export {styles};
